// src/generated/model/getBannerResponseModel.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BannerInfoModel } from './bannerInfoModel';


export interface GetBannerResponseModel { 
    size?: number;
    page?: number;
    totalPages?: number;
    totalElements?: number;
    bannersList?: Array<BannerInfoModel>;
}